Add special function to get prompt inputs embeds
To remove duplicated code a new `get_prompt_inputs_embeds()` method has
been added to the `PeftModel` class to, given the `input_ids` (and
optionally `inputs_embeds`), return the correct `inputs_embeds` with the
corresponding virtual prompts in it (embeddings). In addition, having
this functionality as a separated method allow user to get the
"ready-to-use" input embeddings from outside for further changes before
feeding it to the model.

For example:

>>> inputs_embeds = peft_model.get_prompt_inputs_embeds(
>>>     input_ids=input_ids,
>>>     inputs_embeds=inputs_embeds
>>> )
>>> # here do some other extra magic with `inputs_embeds`
>>> # before calling the base model with it:
>>> output = peft_model.base_model(inputs_embeds=inputs_embeds)